Refactored into separate files
